feat(frontend): static export (SPA) for S3 + CloudFront hosting
The app is fully client-rendered (JWT in localStorage, all data via apiFetch),
so switch Next to `output: 'export'` and host the static bundle on S3+CloudFront
instead of Amplify SSR.

- next.config.ts: output: 'export', trailingSlash, images.unoptimized; drop the
  dev rewrites (they don't run in export — prod routing is NEXT_PUBLIC_API_BASE_URL).
- projects/[id]: export requires generateStaticParams in a Server Component, but
  the page is 'use client'. Split into a server wrapper (page.tsx, emits a
  throwaway shell) + ProjectDetailClient.tsx (the existing UI, reads id via
  useParams). Real /projects/:id is client-rendered; deep links resolve through
  the CloudFront SPA fallback (404 -> /index.html).

Verified: `next build` exports out/ with index.html + per-route pages +
projects/placeholder shell.
